feat(core,io): Delphi-совместимый MC, выбор метода dogleg, формат ZAK
Реализованы критические задачи для совместимости с PrimeINCL:

Алгоритмы траектории:
- Добавлен MinimumCurvatureIntegral — интегральный метод из unitcalc.pas
- Реализованы все особые случаи (u=u1, a=a1, изменение только зенита/азимута)

Расчёт dogleg:
- Добавлен enum DoglegMethod (Cosine/Sine)
- CalcDLSin реализован для Delphi-совместимости
- Настройка метода в ProcessingSettings.dogleg_method

Формат ZAK:
- Полный импорт файлов заключений (#HEADER, #MEASUREMENTS, #END)
- Автоопределение кодировки CP1251/UTF-8
- Гибкий парсинг колонок и разделителей (;,\t)
- Интеграция в format_registry (автодетекция формата)

Исправления сборки:
- CMake: GTK4_CFLAGS_OTHER через target_compile_options
- Устранены все предупреждения компилятора (-Wconversion и др.)

Новые файлы:
- src/io/zak_reader.hpp, src/io/zak_reader.cpp

Изменённые файлы:
- src/model/types.hpp (MinimumCurvatureIntegral, DoglegMethod)
- src/core/trajectory.hpp, trajectory.cpp
- src/io/format_registry.hpp, format_registry.cpp
- CMakeLists.txt
- docs/COMPATIBILITY_MATRIX.md, REPORT.md

Тесты: 19/19 пройдено

+++ src/core/processing.cpp
@@ -171,7 +171,7 @@ WellResult processWell(
     // Обработка остальных точек
     for (size_t i = 1; i < n; ++i) {
         if (i % 100 == 0) {
-            reportProgress(0.1 + 0.6 * static_cast<double>(i) / n, "Расчёт траектории...");
+            reportProgress(0.1 + 0.6 * static_cast<double>(i) / static_cast<double>(n), "Расчёт траектории...");
         }
 
         const auto& prev = data.measurements[i - 1];
